This fix corrects the issue described below,
The `y` version in release could pass through with a change before this.
This corrects it. `y` should not be changed in release. This bug was
noticed in the package flowAI as noticed by @vobencha.

This fix also cleans up the way the if statement is written.

More details about this fix are in the github repository,

https://github.com/Bioconductor/bioc_git_transition/commit/2dedcf3db3fdd9619cf4f7d75ee39657d5e55aad