Hook to prevent merge conflict markers
This pre-receive hook prevents pushes to thg bioconductor git repo,
if there are merge conflict markers preset in the commit.

Users would have to remove the merge conflicts to make sure that they
can push to the git.bioconductor.org repository again. The hook has been
tested against text files, but does not check for binary files.

The prevent_merge_markers hook is the FIRST line of defence for
packages. If there are merge markers the Bioc git server prevents the
entire commit, and not just the files which are affected.